Things I added to take a video of my project to send to a company (Daitonn)
For the SDP project i added lots of details. but for the repair purpose I had to uninstall everything. So, added products we not there after reinstallation. Again I had to take a video of my project to send it to an internship company. That is added in this commit.